# api/auth/swagger.yaml
basePath: /
definitions:
...
  authsdk.ErrorResponse:
    properties:
      error:
        description: Error is the OAuth2 error code (e.g., "invalid_request", "invalid_grant")
        type: string
      error_description:
        description: ErrorDescription is a human-readable description of the error
        type: string
    type: object
...
  authsdk.TokenResponse:
    properties:
      access_token:
        description: AccessToken is the JWT access token used to authenticate API
          requests
        type: string
      expires_in:
        description: ExpiresIn is the lifetime in seconds of the access token
        type: integer
      refresh_token:
        description: RefreshToken is the opaque refresh token used to obtain new access
          tokens
        type: string
      scope:
        description: Scope is the space-delimited list of scopes granted to this token
        type: string
      token_type:
        description: TokenType is always "Bearer" per OAuth2 spec
        type: string
    type: object
...
  http.IntrospectionResponse:
    properties:
      active:
        type: boolean
      amr:
        items:
          type: string
        type: array
      aud:
        items:
          type: string
        type: array
      client_id:
        type: string
      exp:
        type: integer
      iat:
        type: integer
      iss:
        type: string
      jti:
        type: string
      nbf:
        type: integer
      preferred_name:
        type: string
      scope:
        description: Optional fields (only present when active=true)
        type: string
      sid:
        type: string
      sub:
        type: string
      token_type:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: AussieBroadWAN Team
    url: https://github.com/aussiebroadwan/bartab
  description: |-
    Attempting to be OAuth2-compliant for our authentication service providing token management with JWT-based access tokens.

    All tokens are signed using RS256 (RSA-SHA256) and can be verified using the JWKS endpoint.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: BarTab Authentication Service API
  version: 0.1.0
paths:
...
  /v1/oauth2/introspect:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Introspects a token and returns metadata about it (RFC 7662)
      parameters:
      - description: The token to introspect
        in: formData
        name: token
        required: true
        type: string
      - description: Hint about token type (currently only 'access_token' is supported)
        enum:
        - access_token
        in: formData
        name: token_type_hint
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token introspection result
          headers:
            Cache-Control:
              description: no-store
              type: string
            Pragma:
              description: no-cache
              type: string
          schema:
            $ref: '#/definitions/http.IntrospectionResponse'
        "400":
          description: error, error_description
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: error, error_description
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: OAuth2 Token Introspection Endpoint
      tags:
      - OAuth2
  /v1/oauth2/revoke:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: |-
        Revokes a previously issued token (RFC 7009)
        Currently supports revoking refresh tokens only. Access tokens expire naturally.
        The endpoint is idempotent and returns 200 OK even for invalid/unknown tokens to prevent token scanning attacks.
      parameters:
      - description: The token to revoke
        in: formData
        name: token
        required: true
        type: string
      - description: Hint about token type
        enum:
        - access_token
        - refresh_token
        in: formData
        name: token_type_hint
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token revoked successfully (or was already invalid)
          headers:
            Cache-Control:
              description: no-store
              type: string
            Pragma:
              description: no-cache
              type: string
        "400":
          description: error, error_description
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: error, error_description
          schema:
            additionalProperties:
              type: string
            type: object
      summary: OAuth2 Token Revocation Endpoint
      tags:
      - OAuth2
  /v1/oauth2/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Issues access and refresh tokens using OAuth2 grant types (authorization_code,
        refresh_token, client_credentials, mfa_otp).
      parameters:
      - description: Grant type
        enum:
        - authorization_code
        - refresh_token
        - client_credentials
        - mfa_otp
        in: formData
        name: grant_type
        required: true
        type: string
      - description: Authorization code (required for authorization_code grant)
        in: formData
        name: code
        type: string
      - description: Redirect URI (required for authorization_code grant)
        in: formData
        name: redirect_uri
        type: string
      - description: PKCE code_verifier (required when PKCE was used)
        in: formData
        name: code_verifier
        type: string
      - description: Refresh token (required for refresh_token grant)
        in: formData
        name: refresh_token
        type: string
      - description: Client identifier (required for all grants)
        in: formData
        name: client_id
        required: true
        type: string
      - description: Client secret (required for confidential clients)
        in: formData
        name: client_secret
        type: string
      - description: Space-delimited list of scopes
        in: formData
        name: scope
        type: string
      - description: MFA token (required for mfa_otp grant)
        in: formData
        name: mfa_token
        type: string
      - description: MFA method (required for mfa_otp grant) - totp or backup_codes
        in: formData
        name: method
        type: string
      - description: OTP code (required for mfa_otp grant)
        in: formData
        name: otp_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: access_token, refresh_token, token_type, expires_in, scope
          headers:
            Cache-Control:
              description: no-store
              type: string
            Pragma:
              description: no-cache
              type: string
          schema:
            $ref: '#/definitions/authsdk.TokenResponse'
        "400":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "405":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      summary: OAuth2 Token Endpoint
      tags:
      - OAuth2
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"